Move non-construct code out of constructs
