+++ src/constructs/lambda/index.ts
@@ -1,2 +1,1 @@
 export * from "./lambda";
-export * from "./event-sources";
